+++ DataViewer/DataViewer.csproj.user
@@ -7,6 +7,9 @@
     <Compile Update="CultureLandscapeForm.cs">
       <SubType>Form</SubType>
     </Compile>
+    <Compile Update="LandscapeForm.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Update="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
